修改了环境文件、添加copyright、删除中文注释、将sp有关代码放到ulysses中
